schema_registry: shift maybe_authorize_request into service
